Ya está hecho lo de la foto en juegos

+++ Controlador/index_usuarios.php
@@ -235,31 +235,48 @@
 
 
     function modificarJuego(){
-
         //En caso de que se haya modificado la imagen, se sube la nueva img y se borra la antigua, al contrario, se deja la misma imagen que había
+        if(!empty($_FILES["img"]['tmp_name'])){
+            //Guardar imagen en la carpeta para que luego se pueda subir esa ruta a la bd a la hora de modificar
+            $destino=guardarImg();
 
+            require_once("../Modelo/class_juego.php");
+            $juego=new Juego();
 
-        //Guardar imagen en la carpeta para que luego se pueda subir esa ruta a la bd a la hora de modificar
-        $destino=guardarImg();
+            //Coger la ruta de la img actual del juego para luego eliminarla
+            $rutaImgActual=$juego->obtenerImgActual($_POST["idJuego"]);
 
-        require_once("../Modelo/class_juego.php");
-        $juego=new Juego();
-
-        //Coger la ruta de la img actual del juego para luego eliminarla
-        $rutaImgActual=$juego->obtenerImgActual($_POST["idJuego"]);
-
-        //Actualizar con los nuevos datos
-        $modificado=$juego->modificar_juego($_POST["tit"],$_POST["plat"],$_POST["lanz"],$destino,$_POST["idJuego"]);
-        
-        //Borrar la img anterior de la carpeta
-        unlink($rutaImgActual);
-        
-        if($modificado){
-            //Redirigir al menu de juegos y mostrar toast de Exito
-            juegos();
+            //Actualizar con los nuevos datos
+            $modificado=$juego->modificar_juego($_POST["tit"],$_POST["plat"],$_POST["lanz"],$destino,$_POST["idJuego"]);
+            
+            //Borrar la img anterior de la carpeta
+            unlink($rutaImgActual);
+            
+            if($modificado){
+                //Redirigir al menu de juegos y mostrar toast de Exito
+                juegos();
+                exit;
+            }else{
+                //Redirigir al menu de juegos
+                juegos();
+            }
         }else{
-            //Mostrar mensaje de error
+            require_once("../Modelo/class_juego.php");
+            $juego=new Juego();
+
+            $rutaImgActual=$juego->obtenerImgActual($_POST["idJuego"]);
+            $modificado=$juego->modificar_juego($_POST["tit"],$_POST["plat"],$_POST["lanz"],$rutaImgActual,$_POST["idJuego"]);
+
+            if($modificado){
+                //Redirigir al menu de juegos y mostrar toast de Exito
+                juegos();
+                exit;
+            }else{
+                juegos();
+            }
+
         }
+        
     }
